Splitter context for Oppgave og OppgaveFilter for å kunne gjenbruke oppgaver i personoversikten
