+++ src/frontend/context/OppgaveContext.ts
@@ -1,10 +1,8 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
 
 import constate from 'constate';
 
 import { useApp } from './AppContext';
-import { hentLagretOppgaveRequest } from '../Sider/Oppgavebenk/filter/oppgavefilterStorage';
-import { defaultOppgaveRequest } from '../Sider/Oppgavebenk/oppgaverequestUtil';
 import { Oppgave, OppgaveRequest, OppgaverResponse } from '../Sider/Oppgavebenk/typer/oppgave';
 import {
     byggHenterRessurs,
@@ -14,20 +12,13 @@ import {
     RessursStatus,
     RessursSuksess,
 } from '../typer/ressurs';
-import { harStrengtFortroligRolle } from '../utils/roller';
 
 export const [OppgaveProvider, useOppgave] = constate(() => {
-    const { request, saksbehandler, appEnv } = useApp();
+    const { request } = useApp();
     const [oppgaveRessurs, settOppgaveRessurs] =
         useState<Ressurs<OppgaverResponse>>(byggTomRessurs());
     const [laster, settLaster] = useState<boolean>(false);
-    const [lasterOppgaveRequestFraLocaleStorage, settLasterOppgaveRequestFraLocaleStorage] =
-        useState<boolean>(true);
     const [feilmelding, settFeilmelding] = useState<string>();
-
-    const [oppgaveRequest, settOppgaveRequest] = useState<OppgaveRequest>(defaultOppgaveRequest);
-
-    const harSaksbehandlerStrengtFortroligRolle = harStrengtFortroligRolle(appEnv, saksbehandler);
 
     const hentOppgaver = useCallback(
         (data: OppgaveRequest) => {
@@ -38,16 +29,6 @@ export const [OppgaveProvider, useOppgave] = constate(() => {
         },
         [request]
     );
-
-    useEffect(() => {
-        const lagretFiltrering = hentLagretOppgaveRequest(
-            saksbehandler,
-            harSaksbehandlerStrengtFortroligRolle
-        );
-        settOppgaveRequest(lagretFiltrering);
-        settLasterOppgaveRequestFraLocaleStorage(false);
-        hentOppgaver(lagretFiltrering);
-    }, [hentOppgaver, harSaksbehandlerStrengtFortroligRolle, saksbehandler]);
 
     const oppdaterOppgaveEtterTilbakestilling = (oppdatertOppgave: Oppgave) => {
         settOppgaveRessurs((prevState) => {
@@ -121,8 +102,5 @@ export const [OppgaveProvider, useOppgave] = constate(() => {
         tilbakestillFordeling,
         settOppgaveTilSaksbehandler,
         oppdaterOppgaveEtterTilbakestilling,
-        lasterOppgaveRequestFraLocaleStorage,
-        oppgaveRequest,
-        settOppgaveRequest,
     };
 });
